+++ app/routes.php
@@ -8,15 +8,17 @@
  *      Route::addController('mysecondroute', new MyController());
  */
 
-use App\Http\Controllers\LoginController;
-use App\Http\Controllers\LogoutController;
 use App\Http\Controllers\ProductController;
+use App\Http\Controllers\User\ChangePasswordController;
+use App\Http\Controllers\User\LoginController;
+use App\Http\Controllers\User\LogoutController;
 use App\Http\Controllers\UserController;
 use Framework\Authentication\Auth;
 use Framework\Cli\Controllers\WebCliController;
+use Framework\Facades\Http;
 use Framework\Routing\Router;
 
-Router::addFn('', fn() => view('home'));
+    Router::addFn('', fn() => view('home'));
 Router::addFn('imprint', fn() => view('imprint'));
 
 Router::addFn('login', fn() => view('login'));
@@ -25,6 +27,8 @@ Router::addController('login', new LoginController(), 'POST');
 if (Auth::isLoggedIn()) {
     Router::addFn('about', fn() => view('about'));
 
+    Router::addFn('changePassword', fn() => view('entities.user.changePassword'));
+    Router::addController('changePassword', new ChangePasswordController(), 'POST');
     Router::addController('logout', new LogoutController());
 
     Router::addController('cli', new WebCliController());
@@ -36,4 +40,4 @@ if (Auth::isLoggedIn()) {
 
 // TODO only if is admin
 
-Router::pathNotFound(function ($path) { echo 'cannot find path "' . $path. '"'; });
+Router::pathNotFound(fn() => Http::redirect('/'));