+++ CHANGELOG.md
@@ -42,6 +42,7 @@ Types of changes: `Added`, `Changed`, `Deprecate`, `Removed`, `Fixed`, `Secruity
 - Added table for roles and user-role-assignment
 - Added message class to pass messages from controller to view
 - Added session class to set and get values in or from session
+- Added form and controller to change the password
 
 ### Changed
 - Improved the autoloader to work for the new folder structure of version 2